+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 ### 🐛Fixed
 - Error when checking the number of streamlines in the weighted LASSO function
 - Error in verifying voxel correspondence between the lesion mask passed through the dictionary and the ISO weights map
+- Error when checking the number of streamlines in the grouping structure for group LASSO regularization
+- Added check on empty groups when setting group LASSO regularization
 - Tests fail using numpy >= 2.0.0
 
 ---